requirements.txt
alabaster==0.7.12
analytics-python==1.2.9
argh==0.26.2
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
Babel==2.6.0
bcrypt==3.1.7
boto==2.49.0
boto3==1.9.71
botocore==1.12.71
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coverage==4.5.2
cryptography==2.7
Django==2.1.5
django-appconf==1.0.3
django-background-tasks==1.2.0
django-bootstrap4==0.0.7
django-bower==5.2.0
django-common-helpers==0.9.2
django-compat==1.0.15
django-compressor==2.2
django-cron==0.5.1
django-debug-toolbar==1.11
django-easy-pdf==0.1.1
django-encrypted-model-fields==0.5.8
django-extensions==2.1.4
django-extra-views==0.12.0
django-filter==2.0.0
django-formtools==2.1
django-health-check==3.8.0
django-inline-svg==0.1.1
django-ipware==2.1.0
django-libsass==0.7
django-maintenance-mode==0.13.0
django-polymorphic==2.0.3
django-ses==0.8.7
django-session-security==2.6.2
django-simple-history==2.7.0
django-storages==1.7.1
django-tables2==2.0.3
django-widget-tweaks==1.4.3
djangorestframework==3.11.0
docutils==0.14
flake8==3.6.0
future==0.17.1
html5lib==1.0.1
idna==2.8
imagesize==1.1.0
isort==4.3.21
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
lazy-object-proxy==1.4.1
libsass==0.19.1
livereload==2.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.3.14
numpy==1.16.4
packaging==19.0
pandas==0.23.4
pathtools==0.1.2
pbr==5.2.0
Pillow==6.0.0
port-for==0.3.1
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.4.0
pylint==2.3.1
pyparsing==2.4.0
PyPDF2==1.26.0
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
PyYAML==5.1
rcssmin==1.0.6
reportlab==3.5.23
requests==2.14.2
rjsmin==1.0.12
s3transfer==0.1.13
selenium==3.13.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.3
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
sqlparse==0.3.0
stevedore==1.30.1
tornado==6.0.2
typed-ast==1.4.0
unittest-xml-reporting==2.2.1
urllib3==1.24.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
watchdog==0.9.0
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
yapf==0.25.0
